log crdb sql too
