+++ internal/datastore/common/sql.go
@@ -2,8 +2,10 @@ package common
 
 import (
 	"context"
+	"fmt"
 	"math"
 	"runtime"
+	"strings"
 
 	sq "github.com/Masterminds/squirrel"
 	v1 "github.com/authzed/authzed-go/proto/authzed/api/v1"
@@ -376,6 +378,20 @@ func (tqs TupleQuerySplitter) SplitAndExecuteQuery(
 	iter := datastore.NewSliceRelationshipIterator(tuples)
 	runtime.SetFinalizer(iter, datastore.MustIteratorBeClosed)
 	return iter, nil
+}
+
+func InlineSqlArgs(sqlQuery string, args []interface{}) string {
+	for _, arg := range args {
+		var formattedArg string
+		switch arg.(type) {
+		case string:
+			formattedArg = fmt.Sprintf("'%v'", arg)
+		default:
+			formattedArg = fmt.Sprint(arg)
+		}
+		sqlQuery = strings.Replace(sqlQuery, "?", formattedArg, 1)
+	}
+	return sqlQuery
 }
 
 // ExecuteQueryFunc is a function that can be used to execute a single rendered SQL query.
